% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onglets}
\alias{onglets}
\title{Translation of tabs}
\format{
data.frame with 4 columns and as many lines as tabs
\describe{
\item{id}{shinydashboard::menuItem() tab name}
\item{icon}{FontAwesome icon name}
\item{fr}{tab text in French}
\item{en}{tab text in English}
}
}
\usage{
onglets
}
\description{
Translation of tabs
}
\keyword{datasets}
