% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par}
\alias{par}
\title{Barycentres of the known citrus plots in Reunion}
\format{
data.frame 782x3
\describe{
\item{ID}{Plot identifier}
\item{longitude, latitude}{WGS84 coordinates of the plot barycentre}
}
}
\usage{
par
}
\description{
Barycentres of the known citrus plots in Reunion
}
\keyword{datasets}
