% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pluvio}
\alias{pluvio}
\title{polygon simple feature collection containing the median annual rainfall levels of Reunion island from 1986 to 2016}
\format{
sf with 4293 zones and 1 field
\describe{
\item{r_median}{median rainfall}
}
}
\usage{
pluvio
}
\description{
polygon simple feature collection containing the median annual rainfall levels of Reunion island from 1986 to 2016
}
\keyword{datasets}
