% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prelev}
\alias{prelev}
\title{Raw data of the presence of the disease for the barplot}
\format{
data.frame 267x7
\describe{
\item{Id}{Plot identifier}
\item{X, Y}{WGS84 coordinates of the plot barycentre}
\item{Surface}{Area of the plot in square meters}
\item{Altitude}{Elevation of the plot}
\item{Maladie}{Presence (1) or absence (0) of the disease}
\item{Date}{Date of the observation}
}
}
\usage{
prelev
}
\description{
Raw data of the presence of the disease for the barplot
}
\keyword{datasets}
