% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surface_agrume}
\alias{surface_agrume}
\title{Dataframe with the total citrus plot area per commune}
\format{
data.frame 25x2
\describe{
\item{Surface_tot}{Citrus plot area in square meters}
\item{COMMUNE}{Municipality name}
}
}
\usage{
surface_agrume
}
\description{
Dataframe with the total citrus plot area per commune
}
\keyword{datasets}
