% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textesUI}
\alias{textesUI}
\title{Translation of the rest of the user interface}
\format{
data.frame with 3 columns and as many lines as texts to translate
\describe{
\item{id}{unique text identifier}
\item{fr}{text in French}
\item{en}{text in English}
}
}
\usage{
textesUI
}
\description{
Translation of the rest of the user interface
}
\keyword{datasets}
