% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes}
\alias{communes}
\title{polygon simple feature collection containing the municipalities of Reunion island}
\format{
sf with 24 municipalities and 1 field
\describe{
\item{COMMUNE}{Municipality name}
}
}
\usage{
communes
}
\description{
polygon simple feature collection containing the municipalities of Reunion island
}
\keyword{datasets}
