% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlu}
\alias{hlu}
\title{Raw data of the presence of the disease for the interpolation}
\format{
data.frame 616x4
\describe{
\item{id}{Plot identifier}
\item{longitude, latitude}{WGS84 coordinates of the plot barycentre}
\item{date}{Date of the observation}
}
}
\usage{
hlu
}
\description{
Raw data of the presence of the disease for the interpolation
}
\keyword{datasets}
