% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{textesUI}
\alias{textesUI}
\title{Translation of the rest of the user interface}
\format{
data.frame 48x4
}
\usage{
textesUI
}
\description{
Translation of the rest of the user interface
}
\keyword{datasets}
